tests/requirements.txt
pytest~=8.2.0
pytest-cov~=5.0.0
approvaltest~=12.0.0
pytest-approvaltests~=0.2.4
